refactor(core): avanzar baseline canónica y retirar modelo financiero legado
Reemplaza el historial Flyway V1-V060 por una única V1 canónica preliminar y alinea el modelo JPA con PostgreSQL mediante IDs BIGINT, importes NUMERIC y relaciones explícitas.

Introduce cargos, pagos, aplicaciones, movimientos de caja, crédito, stock y outbox; elimina DetallePago, clones financieros, ProcesoEjecutado y consumidores heredados asociados.

Implementa pagos parciales mediante aplicaciones, validación de sobreaplicación, idempotencia, locks ordenados, reversión compensatoria y crédito respaldado por movimientos.

Agrega contratos PostgreSQL para Flyway V1, Hibernate validate, catálogo, tipos monetarios, índices de FK y ausencia de cascadas históricas destructivas.

Actualiza parcialmente backend y frontend al nuevo contrato, elimina pantallas y clientes de DetallePago y transporta importes financieros nuevos como strings decimales.

Agrega documentación del modelo canónico, matriz de relaciones, plan de reset y política de finales de línea mediante .gitattributes.

Checkpoint incompleto: compilación productiva backend y gate estructural PostgreSQL PASS; quedan pendientes frontend completo, clean verify global, seguridad integral, side effects, CI, Docker, documentación final y búsquedas de cierre.

+++ backend/src/main/java/ledance/dto/disciplina/request/DisciplinaHorarioModificacionRequest.java
@@ -1,12 +1,9 @@
-// DisciplinaHorarioModificacionRequest.java
 package ledance.dto.disciplina.request;
 
 import ledance.entidades.DiaSemana;
+
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
-public record DisciplinaHorarioModificacionRequest(
-        Long id,           // Sera null para nuevos horarios
-        DiaSemana diaSemana,
-        LocalTime horarioInicio,
-        Double duracion
-) { }
+public record DisciplinaHorarioModificacionRequest(Long id, DiaSemana diaSemana, LocalTime horarioInicio, BigDecimal duracion) {
+}
