+++ backend/src/main/java/ledance/dto/disciplina/request/DisciplinaHorarioRequest.java
@@ -1,10 +1,9 @@
 package ledance.dto.disciplina.request;
 
 import ledance.entidades.DiaSemana;
+
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
-public record DisciplinaHorarioRequest(
-        DiaSemana diaSemana,
-        LocalTime horarioInicio,
-        Double duracion
-) {}
+public record DisciplinaHorarioRequest(DiaSemana diaSemana, LocalTime horarioInicio, BigDecimal duracion) {
+}
