+++ backend/src/main/java/ledance/dto/disciplina/response/DisciplinaHorarioResponse.java
@@ -1,11 +1,9 @@
 package ledance.dto.disciplina.response;
 
 import ledance.entidades.DiaSemana;
+
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
-public record DisciplinaHorarioResponse(
-        Long id,
-        DiaSemana diaSemana,
-        LocalTime horarioInicio,
-        Double duracion
-) {}
+public record DisciplinaHorarioResponse(Long id, DiaSemana diaSemana, LocalTime horarioInicio, BigDecimal duracion) {
+}
